Changed below solar generation automation from notification to counting days.
If it's below expected (set by input number) then it will incremement the number of consecutive days it's below expectation.

If the consecutive days is over 6 i.e a week (for now), then it will send a notification. This will account for odd days that are below.

I expect this to trigger more during the winter months where I will need to change the inverter charging strategy / priority.

Update #54