Changed motioned detected which cancel timers from restart to queued mode.
There is a tiny chance if there's constant motion it will never complete the automation in time. For example if it cancels a dim light timner and restarts constantly due to motion triggers and never hit the cancel turn off light timer. Shorter the timer, the higher likely hood this scenario could happen.

Updated #62.