Added check for how long a person has been home.
This mitigates an edge case where someone gets home but still in the car at the point when the automation runs. Due to the proximity, Home Assistant might see them as "home" but still outside causing the alarm to turn on.

To mitigate this, people of interest must be home for at least 5 minutes before the alarm will turn on. This gives the person 5 minutes outside the home to get in.

By that point, the closing of the door will automation would turn on the alarm as well.