Updated Jinja call to calculate charging rate with the max charge rate parameter.
Created timer to check for excess solar. This can happen where the
excess remains above threshold when it's charging and it will never
trigger the automation for excess solar energy again, meaning the
EcoFlow could increase it's charging rate to soak up the excess but
never does. It's not a full proof way and until I figure out how to
trigger the automation, this will hopefully reduce the amount of wasted
solar energy.

Update #94