Added EcoFlow notification when plug is switched off.
A smart plug controls the power going to EcoFlow because it continues to pull power even when the charge level is above the backup energy reserve.

Added scenario where import rate goes above 0 but still below the export rate. In this case, it will continue to run from the mains.

Added notification when EcoFlow battery level is low and plug is switched off.

Added notification when rates fall and due to be set to charge however the plug has been turned off.