Added missing tap dial action which was fast rotate.
Changed tap dial step by double. The value it was increasing the brightness was too slow.

Added a transition time. Gives it a better touch and slightly covers the delay when using the remote.

Added value chedk to see if it maxes out the brightness.

Changed automation to be queued rather than signle.