Added target rate.
A target rate can be manually adjusted and I see this as being changed more frequently compared to below export rate.

This will allow periods where it's below export rate but only charge when it is significantly less or delay charging to a more favourable period semi automatically.

For example:
1. 14:00-15:00: 14p/kWh
2. 15:00-15:30: 9p/kWh
3. 15:30-16:00: 13p/kWh

You can use it to target the 9p/kWh or include 13p but not 14p without having to constantly adjust schedules. I can see these would differ in frequency of usage between summer and winter. i.e Winter will be charging as much from the grid due to lack of sun.

Update #102