Perform a default check on import rate check. This removes the required field and default to the current rate if no value is passed.
Fix using event in nurmeric state condition for rate checks. Event sensors do not work with numeric state conditons and triggers. Swapped these out for template conditions on the new pattern.

Added a cost 0p/kWh condition. Although this is the same as negative rates, this allows adjustments in the future to cater for this + for testing purposes.

Removed export field to check office air conditioner.

Update #108